+++ Cite.Api/Controllers/ScoringModelController.cs
@@ -3,13 +3,17 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Cite.Api.Data.Enumerations;
+using Cite.Api.Infrastructure.Authorization;
 using Cite.Api.Infrastructure.Extensions;
 using Cite.Api.Infrastructure.Exceptions;
+using Cite.Api.Infrastructure.Identity;
 using Cite.Api.Infrastructure.QueryParameters;
 using Cite.Api.Services;
 using Cite.Api.ViewModels;
@@ -20,12 +24,14 @@ namespace Cite.Api.Controllers
     public class ScoringModelController : BaseController
     {
         private readonly IScoringModelService _scoringModelService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly ICiteAuthorizationService _authorizationService;
+        private readonly IIdentityResolver _identityResolver;
 
-        public ScoringModelController(IScoringModelService scoringModelService, IAuthorizationService authorizationService)
+        public ScoringModelController(IScoringModelService scoringModelService, ICiteAuthorizationService authorizationService, IIdentityResolver identityResolver)
         {
             _scoringModelService = scoringModelService;
             _authorizationService = authorizationService;
+            _identityResolver = identityResolver;
         }
 
         /// <summary>
@@ -42,7 +48,21 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "getScoringModels")]
         public async Task<IActionResult> Get([FromQuery] ScoringModelGet queryParameters, CancellationToken ct)
         {
-            var list = await _scoringModelService.GetAsync(queryParameters, ct);
+            var userIdString = _identityResolver.GetId().ToString();
+            var list = new List<ScoringModel>();
+            // get ALL scenario templates
+            if (await _authorizationService.AuthorizeAsync([SystemPermission.ViewScoringModels], ct))
+            {
+                list = (await _scoringModelService.GetAsync(queryParameters, ct)).ToList();
+            }
+            // get scenario templates the user can access
+            else if (queryParameters.UserId == null || queryParameters.UserId == userIdString)
+            {
+                queryParameters.UserId = userIdString;
+                list = (await _scoringModelService.GetAsync(queryParameters, ct)).ToList();
+            }
+            AddPermissions(list);
+
             return Ok(list);
         }
 
@@ -60,10 +80,15 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "getScoringModel")]
         public async Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<ScoringModel>(id, [SystemPermission.ViewScoringModels], [ScoringModelPermission.ViewScoringModel], ct))
+                throw new ForbiddenException();
+
             var scoringModel = await _scoringModelService.GetAsync(id, ct);
 
             if (scoringModel == null)
                 throw new EntityNotFoundException<ScoringModel>();
+
+            AddPermissions(scoringModel);
 
             return Ok(scoringModel);
         }
@@ -83,8 +108,13 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "createScoringModel")]
         public async Task<IActionResult> Create([FromBody] ScoringModel scoringModel, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScoringModels], ct))
+                throw new ForbiddenException();
+
             scoringModel.CreatedBy = User.GetId();
             var createdScoringModel = await _scoringModelService.CreateAsync(scoringModel, ct);
+            AddPermissions(createdScoringModel);
+
             return CreatedAtAction(nameof(this.Get), new { id = createdScoringModel.Id }, createdScoringModel);
         }
 
@@ -103,7 +133,13 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "copyScoringModel")]
         public async Task<IActionResult> Copy(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScoringModels], ct)
+                || !await _authorizationService.AuthorizeAsync<ScoringModel>(id, [SystemPermission.ViewScoringModels], [ScoringModelPermission.ViewScoringModel], ct))
+                throw new ForbiddenException();
+
             var createdScoringModel = await _scoringModelService.CopyAsync(id, ct);
+            AddPermissions(createdScoringModel);
+
             return CreatedAtAction(nameof(this.Get), new { id = createdScoringModel.Id }, createdScoringModel);
         }
 
@@ -124,8 +160,13 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "updateScoringModel")]
         public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] ScoringModel scoringModel, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<ScoringModel>(id, [SystemPermission.EditScoringModels], [ScoringModelPermission.EditScoringModel], ct))
+                throw new ForbiddenException();
+
             scoringModel.ModifiedBy = User.GetId();
             var updatedScoringModel = await _scoringModelService.UpdateAsync(id, scoringModel, ct);
+            AddPermissions(updatedScoringModel);
+
             return Ok(updatedScoringModel);
         }
 
@@ -144,6 +185,9 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "deleteScoringModel")]
         public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<ScoringModel>(id, [SystemPermission.ManageScoringModels], [ScoringModelPermission.ManageScoringModel], ct))
+                throw new ForbiddenException();
+
             await _scoringModelService.DeleteAsync(id, ct);
             return NoContent();
         }
@@ -156,7 +200,12 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "uploadJsonFiles")]
         public async Task<IActionResult> UploadJsonAsync([FromForm] FileForm form, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScoringModels], ct))
+                throw new ForbiddenException();
+
             var result = await _scoringModelService.UploadJsonAsync(form, ct);
+            AddPermissions(result);
+
             return Ok(result);
         }
 
@@ -168,12 +217,27 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "downloadJson")]
         public async Task<IActionResult> DownloadJsonAsync(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<ScoringModel>(id, [SystemPermission.ViewScoringModels], [ScoringModelPermission.ViewScoringModel], ct))
+                throw new ForbiddenException();
+
             (var stream, var fileName) = await _scoringModelService.DownloadJsonAsync(id, ct);
 
             // If this is wrapped in an Ok, it throws an exception
             return File(stream, "application/octet-stream", fileName);
         }
 
+        private void AddPermissions(IEnumerable<ScoringModel> list)
+        {
+            foreach (var item in list)
+            {
+                AddPermissions(item);
+            }
+        }
+
+        private void AddPermissions(ScoringModel item)
+        {
+            item.ScoringModelPermissions = _authorizationService.GetScoringModelPermissions(item.Id).SelectMany(m => m.Permissions).Select(m => m.ToString()).ToList();
+        }
+
     }
 }
-
